Refactor link scanner classes into namespace
